[FEATURE] Added CDN Service and ViewHelper, DocumentHead Utility
This allows single-line inclusion of any version of
jQuery you wish for. Abstract version definitions such
as "1.8" or even "1" simply loads the newest stable release.

The new DocumentHead utility allows advanced access to the data
inserted in additionalHeaderData, most notably the support for
enforcing a load-order to your scripts.

Documentation:
http://fedext.net/api-documentation/services/cdn-service/
http://fedext.net/api-documentation/services/documenthead-service/
http://fedext.net/viewhelper-documentation/resource-assistants/jquery-ui-cdn-loader/



git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/fed/trunk@48411 735d13b6-9817-0410-8766-e36946ffe9aa