[FEATURE] Automatic ExtJS4 Model.js template detection
This enables the detection and usage of custom ExtJS4 Model 
definitions by scanning the path specified as partialRootPath 
in TypoScript for your extension (the extension, not a plugin). 
Changing this path in your Typoscript also makes FED look for 
these template files in that location.

Two files may be used and are detected, in order:

EXT:myext/Resources/Private/Partials/MyDomainObject/Model.js

EXT:myext/Resources/Private/Partials/Model.js

This means that if the first file is not found the second one 
is used and functions as a "master" template for any Domain 
Object which does not have a custom template.

If no template is specified at all the FED default is still used.



git-svn-id: https://svn.typo3.org/TYPO3v4/Extensions/fed/trunk@48494 735d13b6-9817-0410-8766-e36946ffe9aa