[~CLEANUP] A few cleanups in BE rendering for FCEs
Not a breaking change by any means, but if your Preview section
does NOT start with a block element, your preview is assumed to
be text only and will simply be appended to the same line as the
header and FCE type. Also changed tag from a header to simple bold.

Change-Id: I37ee63a1a9d390db63b343750ba6eb902aa9faf3
Fixes: #28864