[+BUGFIX] Missing usage of "as" argument, Page/RenderContentViewHelper
This change fixes a problem with Page/RenderContentViewHelper which
before this change would always use the variable name "content" when
rendering page content records manually inside tag content. ViewHelper
now uses the string specified in the "as" argument as the variable
name to contain content records when doing manual rendering.

Change-Id: I13fa6ecf47958118ad8b67b75e17312b9fbadcc7