[+FEATURE] Backend Record Controller feature
This feature adds automatic calling of Extbase controllers
in a Backend context. The feature relies on Extbase's table
mapping (config.tx_extbase.persistence.classes) to turn a
table name into a Controller class name, then calls methods
on that Controller depending on which action (update, delete,
move, create etc) is being taken. If you have not registered
a table-to-objectType and vice-versa mapping at
config.tx_extbase.persistence.classes then no calls will be
made.

The round-trip through Backend controllers should NOT be
used in combination with Repositories, unless for read-only
purposes. Controller actions take a single Model object as
parameter, lets you modify the properties of that Model object,
then expects the Model object to be returned - after which it
is converted again back into a regular array. The array is then
saved to DB.

Also included are Model classes for ContentElement and Page,
Backend Controllers for ContentElement and Page - along with
some new BE-related support methods in DomainObjectInfo.

This is major change and could cause problems so please test.
The entire hook logic from TCEMain has been changed, which
could cause problems particularly with Fluid CE/Page features.

As this feature is only enabled if a mapping is used and a
Backend Controller exists, there is no switch to enable/disable.

Change-Id: I2c2bc9a81e00b241a93540a3b4ab58edaa99d820
Fixes: #29902