[+FEATURE] MenuViewHelper useShortcutTarget, classFirst and classLast
This change adds "useShortcutUrl" argument which when TRUE
will make shortcuts in menu link to the shortcut's location
rather than the target location.

Adds "classFirst" argument which sets class attribute for
first element and "classLast" which does same but for last
element.

Also added is support for doktype and exclusion of sysfolders
from menu.

Change-Id: I37f216a1da11d38352e5c633088cf6ab5ed96824